refactor: 코드 독립성 향상을 위한 분할
